+++ ts/components/SectionStatusComponent.tsx
@@ -3,13 +3,13 @@ import { connect } from "react-redux";
 import _ from "lodash";
 import { StyleSheet, TouchableWithoutFeedback, View } from "react-native";
 import { Text } from "native-base";
-import { SectionStatus, SectionStatusKey } from "../api/backendPublic";
 import { GlobalState } from "../store/reducers/types";
 import { sectionStatusSelector } from "../store/reducers/backendStatus";
 import I18n from "../i18n";
 import { maybeNotNullyString } from "../utils/strings";
 import { openWebUrl } from "../utils/url";
 import { getLocalePrimary } from "../utils/locale";
+import { SectionStatus, SectionStatusKey } from "../types/backendStatus";
 import { IOColors } from "./core/variables/IOColors";
 import IconFont from "./ui/IconFont";
 import { Label } from "./core/typography/Label";
