+++ ts/components/__tests__/SectionStatusComponent.test.tsx
@@ -3,11 +3,11 @@ import { render, fireEvent } from "@testing-library/react-native";
 import { Provider } from "react-redux";
 import * as React from "react";
 import { some } from "fp-ts/lib/Option";
-import { SectionStatus, SectionStatusKey } from "../../api/backendPublic";
 import SectionStatusComponent from "../SectionStatusComponent";
 import I18n, { setLocale } from "../../i18n";
 import { openWebUrl } from "../../utils/url";
 import { IOColors } from "../core/variables/IOColors";
+import { SectionStatus, SectionStatusKey } from "../../types/backendStatus";
 
 jest.mock("../../utils/url");
