+++ ts/store/actions/backendStatus.ts
@@ -2,7 +2,7 @@
  * Action types and action creator related to BackedServicesStatus.
  */
 import { ActionType, createStandardAction } from "typesafe-actions";
-import { BackendStatus } from "../../api/backendPublic";
+import { BackendStatus } from "../../types/backendStatus";
 
 export const backendStatusLoadSuccess = createStandardAction(
   "BACKEND_STATUS_LOAD_SUCCESS"
